Patch --sizes option. (Was mistakenly calling union_size, and additionally, the HLL overload was incorrectly specified.

+++ src/flatten_all.cpp
@@ -39,13 +39,13 @@ int flatten_all(const std::vector<std::string> &fpaths, const std::string outpat
     if(std::fwrite(&numk, sizeof(numk), 1, ofp) != 1) throw std::runtime_error("Failed to write nk");
     if(std::fwrite(&ne, sizeof(ne), 1, ofp) != 1) throw std::runtime_error("Failed to write num entries");
     std::fwrite(&number_sets, sizeof(number_sets), 1, ofp);
-    std::fprintf(stderr, "Wrote %u nk and %zu num e and %zu num sets\n", numk, ne, size_t(number_sets));
+    std::fprintf(stderr, "Wrote %u nk and %zu num e and %zu num sets\n", numk, size_t(ne), size_t(number_sets));
     if(std::fwrite(k_values.data(), sizeof(unsigned), k_values.size(), ofp) != k_values.size()) throw std::runtime_error("Wrong count of written values");
     size_t nwritten;
     if((nwritten = std::fwrite(outp, sizeof(float), nk * ne, ofp)) != nk * ne) {
         throw std::runtime_error("Failed to write nk * ne");
     }
-    std::fprintf(stderr, "Wrote %zu items (%zu * %zu) [%zu bytes]\n", nk * ne, nk, ne, nk * ne * 4);
+    std::fprintf(stderr, "Wrote %zu items (%zu * %zu) [%zu bytes]\n", size_t(nk * ne), nk, size_t(ne), size_t(nk * ne * 4));
     std::fclose(ofp);
     std::free(outp);
     return 0;
